chore(python): reorder dependency-heavy code
* move Construct helpers to a separate module
* move `ui` into `cli` now that nothing in the core library depends on
it
